Migrate to clean architecture
