+++ lib/src/core/app_router.gr.dart
@@ -6,7 +6,6 @@
 
 // ignore_for_file: type=lint
 // coverage:ignore-file
-
 part of 'app_router.dart';
 
 abstract class _$AppRouter extends RootStackRouter {
