+++ lib/src/features/home_store/presentation/ui/widget/hot_sales_carousel.dart
@@ -1,0 +1,159 @@
+import 'package:flutter/material.dart';
+
+class HotSalesCarousel extends StatefulWidget {
+  const HotSalesCarousel({Key? key, required this.isNew}) : super(key: key);
+
+  final bool isNew;
+
+  @override
+  HotSalesCarouselState createState() => HotSalesCarouselState();
+}
+
+class SalesItem {
+  final String imageUrl;
+  final String pageTitle;
+  final String pageDescription;
+
+  SalesItem({
+    required this.imageUrl,
+    required this.pageTitle,
+    required this.pageDescription,
+  });
+}
+
+class HotSalesCarouselState extends State<HotSalesCarousel> {
+  final List<SalesItem> salesItems = [
+    SalesItem(
+      imageUrl:
+      'https://media.discordapp.net/attachments/988860107423100938/1111314904872386662/image.png',
+      pageTitle: 'Iphone 12',
+      pageDescription: 'Super. Mega. Rapido.',
+    ),
+    SalesItem(
+      imageUrl:
+      'https://media.discordapp.net/attachments/988860107423100938/1111314904872386662/image.png',
+      pageTitle: 'Some other title',
+      pageDescription: 'Some other description',
+    ),
+    SalesItem(
+      imageUrl:
+      'https://media.discordapp.net/attachments/988860107423100938/1111314904872386662/image.png',
+      pageTitle: 'Another title',
+      pageDescription: 'Another description',
+    )
+  ];
+
+  int currentIndex = 0;
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 182,
+      child: PageView.builder(
+        itemCount: salesItems.length,
+        onPageChanged: (index) {
+          setState(() {
+            currentIndex = index;
+          });
+        },
+        itemBuilder: (context, index) {
+          final item = salesItems[index];
+          return Padding(
+            padding: const EdgeInsets.symmetric(horizontal: 16),
+            child: ClipRRect(
+              borderRadius: BorderRadius.circular(8.0),
+              child: Stack(
+                fit: StackFit.expand,
+                children: [
+                  Image.network(
+                    item.imageUrl,
+                    fit: BoxFit.cover,
+                  ),
+                  Positioned(
+                    top: 8,
+                    left: 8,
+                    child: Container(
+                      width: 27,
+                      height: 27,
+                      decoration: const BoxDecoration(
+                        shape: BoxShape.circle,
+                        color: Color(0xFFFF6E4E),
+                      ),
+                      child: const Center(
+                        child: Text(
+                          'New',
+                          style: TextStyle(
+                            color: Colors.white,
+                            fontWeight: FontWeight.bold,
+                            fontSize: 10,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                  Positioned(
+                    top: 60,
+                    left: 8,
+                    child: Container(
+                      padding: const EdgeInsets.symmetric(horizontal: 8),
+                      color: Colors.black.withOpacity(0),
+                      child: Text(
+                        item.pageTitle,
+                        style: const TextStyle(
+                          color: Colors.white,
+                          fontWeight: FontWeight.bold,
+                          fontSize: 25,
+                        ),
+                      ),
+                    ),
+                  ),
+                  Positioned(
+                    top: 90,
+                    left: 8,
+                    child: Container(
+                      padding: const EdgeInsets.symmetric(horizontal: 8),
+                      color: Colors.black.withOpacity(0),
+                      child: Text(
+                        item.pageDescription,
+                        style: const TextStyle(
+                          color: Colors.white,
+                          fontSize: 11,
+                        ),
+                      ),
+                    ),
+                  ),
+                  Positioned(
+                    top: 120,
+                    left: 8,
+                    child: Container(
+                      padding: const EdgeInsets.symmetric(horizontal: 8),
+                      color: Colors.black.withOpacity(0),
+                      child: TextButton(
+                        onPressed: () {},
+                        style: const ButtonStyle(
+                          padding: MaterialStatePropertyAll(
+                            EdgeInsets.symmetric(horizontal: 25),
+                          ),
+                          backgroundColor: MaterialStatePropertyAll(
+                            Colors.white,
+                          ),
+                        ),
+                        child: const Text(
+                          'Buy now!',
+                          style: TextStyle(
+                            color: Color(0xFF010035),
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          );
+        },
+      ),
+    );
+  }
+}
