+++ lib/src/features/home_store/presentation/ui/widget/qr_code_button.dart
@@ -1,0 +1,25 @@
+import 'package:flutter/material.dart';
+
+class QRCodeButton extends StatelessWidget {
+  const QRCodeButton({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return DecoratedBox(
+      decoration: const ShapeDecoration(
+        shape: CircleBorder(),
+        color: Color(0xFFFF6E4E),
+      ),
+      child: IconButton(
+        onPressed: () {},
+        splashRadius: 25,
+        icon: const Icon(Icons.qr_code_outlined),
+        color: Colors.white,
+        tooltip: 'AMONG US',
+        style: ButtonStyle(
+          foregroundColor: MaterialStateProperty.all(const Color(0xFFFF6E4E)),
+        ),
+      ),
+    );
+  }
+}
