+++ pubspec.yaml
@@ -25,12 +25,12 @@ flutter:
   fonts:
     - family: Mark Pro
       fonts:
-        - asset: fonts/MarkPro.ttf
-        - asset: fonts/MarkProBold.ttf
+        - asset: assets/fonts/MarkPro.ttf
+        - asset: assets/fonts/MarkProBold.ttf
           weight: 700
-        - asset: fonts/MarkProMedium.ttf
+        - asset: assets/fonts/MarkProMedium.ttf
           weight: 500
-        - asset: fonts/MarkProHeavy.ttf
+        - asset: assets/fonts/MarkProHeavy.ttf
           weight: 900
